created gallery component added DC trips
