+++ app/components/NavBar.js
@@ -3,6 +3,7 @@
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
+import NavDropdown from 'react-bootstrap/NavDropdown';
 import { basePath } from '@/next.config.mjs';
 
 const NavBar = () => {
@@ -14,6 +15,13 @@ const NavBar = () => {
                     <Nav.Link className='text-white' href={`${basePath}/`}>Home</Nav.Link>
                     <Nav.Link className='text-white' href={`${basePath}/turksandcaicos`}>Turks and Caicos</Nav.Link>
                     <Nav.Link className='text-white' href={`${basePath}/newyork`}>New York</Nav.Link>
+                    <NavDropdown title={
+                        <span className="text-white">Washington DC</span>
+                    } id="dc-nav-dropdown">
+                        <NavDropdown.Item href={`${basePath}/washingtondc2024`}>Cherry Blossom 2024</NavDropdown.Item>
+                        <NavDropdown.Item href={`${basePath}/washingtondc2023`}>Christmas 2023</NavDropdown.Item>
+                        <NavDropdown.Item href={`${basePath}/washingtondcdaytrip`}>Day Trip 2023</NavDropdown.Item>
+                    </NavDropdown>
                 </Nav>
             </Container>
         </Navbar>
