+++ app/washingtondc2023/page.js
@@ -1,0 +1,7 @@
+'use client';
+
+import Gallery from "../components/Gallery";
+
+export default function washingtondc2023() {
+    return <Gallery jsonFilePath="washingtondc2023.json" pageTitle="Christmas 2023" />;
+}
