+++ app/washingtondc2024/page.js
@@ -1,0 +1,7 @@
+'use client';
+
+import Gallery from "../components/Gallery";
+
+export default function washingtondc2024() {
+    return <Gallery jsonFilePath="washingtondc2024.json" pageTitle="Cherry Blossoms 2024" />;
+}
