+++ app/washingtondcdaytrip/page.js
@@ -1,0 +1,7 @@
+'use client';
+
+import Gallery from "../components/Gallery";
+
+export default function washingtondcdaytrip() {
+    return <Gallery jsonFilePath="washingtondcdaytrip.json" pageTitle="Day Trip 2023" />;
+}
