+++ public/washingtondc2023.json
@@ -1,0 +1,27 @@
+[
+  {
+    "src": "/pandj/washingtondc2023/462546785_1555346431751143_1317086577483807359_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondc2023/462551355_1255557458974012_3723954703003357656_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondc2023/462552343_1988291058354953_2883543673184519440_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondc2023/462557531_564848386017148_3946972639069367608_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondc2023/462558894_1260376188612003_3058140515193332133_n.jpg",
+    "title": "",
+    "date": ""
+  }
+]
