+++ public/washingtondcdaytrip.json
@@ -1,0 +1,37 @@
+[
+  {
+    "src": "/pandj/washingtondcdaytrip/1..jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondcdaytrip/462413080_1212091563430716_3843890002332460822_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondcdaytrip/462549392_567554342397390_9187398466114914794_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondcdaytrip/462549872_3773118936233194_4721672980626461852_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondcdaytrip/462550910_827905062594732_2687884133503500101_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondcdaytrip/462552342_1107061174310434_4470185380152505406_n.jpg",
+    "title": "",
+    "date": ""
+  },
+  {
+    "src": "/pandj/washingtondcdaytrip/462552352_1209266747026507_753934134348233993_n.jpg",
+    "title": "",
+    "date": ""
+  }
+]
